packages/app_runner/requirements.txt
python-dotenv==1.1.1
strands-agents==1.3.0
strands-agents-tools==0.2.3
strands-agents-builder==0.1.7
fastapi[standard]==0.116.1
uvicorn[standard]==0.35.0
bacinet==0.0.6